% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datphase1}
\alias{datphase1}
\title{A subset of meta-analytic data set in Michie et al. (2015)}
\format{A data frame of 106 studies with 26 moderators
\itemize{
  \item g: The effect size of each study
  \item vark: The sampling variance of each study
  \item T1 to T26: moderators. Behavior change techniques.
  }}
\usage{
data(datphase1)
}
\description{
Data are heal psychological interventions that used at least one of the
motivation-enchancing behavior change techniques T1, T2, T4, and T25.
}
\keyword{data}
