% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtreeSupport.R
\name{REmrt.fit1}
\alias{REmrt.fit1}
\title{A function to grow the tree}
\usage{
REmrt.fit1(mf, maxL)
}
\arguments{
\item{mf}{the data.frame to grow the tree}

\item{newdata}{the new data set}
}
\value{
a matrix consists of node lables
}
\description{
A function to grow the tree
}
\keyword{internal}
