% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtree.R
\name{REmrt}
\alias{REmrt}
\title{A function to fit random effects RE meta-tree}
\usage{
REmrt(formula, data, vi, c = 1, maxL = 10L, minsplit = 2L, delQ = 0.001,
  n.fold = 10, ...)
}
\arguments{
\item{formula}{A formula, with a response variable (usually the effect size) and the interested moderators but no interaction terms.}

\item{data}{A data frame of a meta-analytic data set, including the effect sizes, sampling variance, and the potential moderators.}

\item{vi}{The column name of the sampling variance in the data set.}

\item{c}{A non-negative scalar.The pruning parameter to prune the initial tree by "c-standard-error" rule.}

\item{maxL}{the maximum number of splits}

\item{minsplit}{the minimal number of studies in a parent node to be split}

\item{delQ}{the stopping rule for decrease of between-subgroups Q. Any split that does not decrease the between-subgroups Q is not attempted.}

\item{n.fold}{the number of folds to perform the cross-validation}

\item{...}{Additional arguments to be passed.}
}
\value{
If no moderator effect is detected, the function will return a list including the following objects:

n: The total number of the studies

Q: The Q-statistics for the heterogeneity test

df: The degree of freedoms of the heterogeneity test

pval.Q: The p-value for the heterogeneity test

g: The overall effect size for all studies

se: The standard error of the overall effect size

zval: The test statistic of the overall effect sie

pval: The p-value for the test statistic of the overall effect size

ci.lb: The lower bound of the confidence interval for the overall effect size

ci.ub: The upper bound of the confidence interval for the overall effect size

call: The matched call

If  moderator effect(s) is(are) detected, the function will return a list including the following objects:

tree: A data frame that represents the tree, with the Q-between and the tau^2 after each split.

n: The number of the studies in each subgroup

moderators: the names of identified moderators

Qb: The between-subgroups Q-statistic

tau2: The estimate for the residual heterogeneity

df: The degree of freedoms of the between-subgroups Q test

pval.Qb: The p-value of the between-subgroups Q test

g: The subgroup overall effect sizes, based on Hedges'g

se: The standard errors of subgroup overall effect sizes

zval: The test statistics of the subgroup overall effect sizes

pval: The p-value for the test statistics of the subgroup overall effect sizes

ci.lb: The lower bounds of the confidence intervals

ci.ub: The upper bounds of the confidence intervals

call: The matched call

cv.res: The cross-validation table

data: the data set subgrouped by the fitted tree
}
\description{
A function to fit random effects RE meta-tree
}
\examples{
data(dat.BCT2015)
REtree <- REmrt(g ~ T1 + T2+ T4 +T25, vi = vi, data = dat.BCT2015, c = 0)
print(REtree)
}
